django_jobs/requirements.txt
Django==1.9.1
django-extensions==1.6.1
django-filter
django-storages-redux==1.3
django-tinymce==2.2.0
djangorestframework==3.3.2
Markdown==2.6.5
psycopg2==2.6.1
six==1.10.0
Werkzeug==0.11.3
django-cors-headers==1.1.0
djangorestframework-jwt==1.7.2
Pillow==3.1.0
boto==2.39.0
drf-extensions==0.2.8
uWSGI==2.0.13.1
raven==5.21.0
-e git+git://github.com/PritishC/django-proxy-storage.git#egg=django-proxy-storage
